--- a/src/Paprika/Utils/ParallelUnbalancedWork.cs
+++ b/src/Paprika/Utils/ParallelUnbalancedWork.cs
@@ -1,3 +1,5 @@
+using System.Runtime.InteropServices;
+
 namespace Paprika.Utils;
 
 /// <summary>
@@ -43,7 +45,10 @@ public class ParallelUnbalancedWork : IThreadPoolWorkItem
         new ParallelUnbalancedWork(data).Execute();
 
         // If there are still active threads, wait for them to complete
-        data.Event.Wait();
+        if (data.ActiveThreads > 0)
+        {
+            data.Event.Wait();
+        }
     }
 
     /// <summary>
@@ -124,13 +129,20 @@ public class ParallelUnbalancedWork : IThreadPoolWorkItem
     /// </summary>
     private class SharedCounter(int fromInclusive)
     {
-        private int _index = fromInclusive;
+        private PaddedValue _index = new(fromInclusive);
 
         /// <summary>
         /// Gets the next index in a thread-safe manner.
         /// </summary>
         /// <returns>The next index.</returns>
-        public int GetNext() => Interlocked.Increment(ref _index) - 1;
+        public int GetNext() => Interlocked.Increment(ref _index.Value) - 1;
+
+        [StructLayout(LayoutKind.Explicit, Size = 128)]
+        private struct PaddedValue(int value)
+        {
+            [FieldOffset(64)]
+            public int Value = value;
+        }
     }
 
     /// <summary>
@@ -228,7 +240,10 @@ public class ParallelUnbalancedWork : IThreadPoolWorkItem
             new InitProcessor<TLocal>(data).Execute();
 
             // If there are still active threads, wait for them to complete
-            data.Event.Wait();
+            if (data.ActiveThreads > 0)
+            {
+                data.Event.Wait();
+            }
         }
 
         /// <summary>