Lower allocating Parallel.For that handles unbalanced work loads better

Rather than partitioning the range and each thread processing that number of items; it creates that number of ThreadPool worker (minus one for main thread); then processes the next available index until the range is complete, meaning one thread may process different numbers of items from another.

This should reduce overall end to end latency also.

Before

| Method                           | count | Mean      | Error    | StdDev   | Gen0   | Code Size | Allocated |
|--------------------------------- |------ |----------:|---------:|---------:|-------:|----------:|----------:|
| Or_BitMapFilter_OfN_128_Multiple | 16    |  24.82 us | 0.346 us | 0.289 us | 0.3662 |   2,464 B |   6.19 KB |
| Or_BitMapFilter_OfN_128_Multiple | 32    |  41.07 us | 0.807 us | 1.232 us | 0.3662 |   2,464 B |   6.21 KB |
| Or_BitMapFilter_OfN_128_Multiple | 64    | 221.41 us | 3.831 us | 2.991 us | 0.2441 |   2,447 B |   7.58 KB |

After

| Method                           | count | Mean      | Error    | StdDev   | Gen0   | Code Size | Allocated |
|--------------------------------- |------ |----------:|---------:|---------:|-------:|----------:|----------:|
| Or_BitMapFilter_OfN_128_Multiple | 16    |  27.71 us | 0.095 us | 0.079 us | 0.0610 |   1,126 B |   1.20 KB |
| Or_BitMapFilter_OfN_128_Multiple | 32    |  34.58 us | 0.557 us | 0.521 us | 0.0610 |   1,126 B |   1.45 KB |
| Or_BitMapFilter_OfN_128_Multiple | 64    | 180.98 us | 3.559 us | 6.139 us |      - |   1,126 B |   1.95 KB |